requirements.txt
arxiv~=2.0.0
pandas~=2.0.3
tqdm~=4.66.1
openai~=1.12.0
mistralai~=0.0.12
python-dotenv~=1.0.1

# For the benchmark and plots:
scikit-learn~=1.3.2
seaborn~=0.13.2
matplotlib~=3.7.5